refactor(model): 당첨 종류를 검증하는 기능 분리
enum을 사용하여 당첨 금액 및 해야할 일을 선언합니다.
